feat: regroup admin commands under /jd4h-admin, /rngdle-admin, /admin
